+++ roles/supervisor/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Use unified exclusive template lookup
+- Deprecate environment oriented templates
+- Deprecate dict's array configs
 
 ## [2.0.2] - 2020-02-13
 ### Added